fastapi/requirements.txt
# FastAPI 및 기본 의존성
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6  # Form 데이터 및 파일 업로드 처리

# 데이터베이스 (ddri_db)
pymysql>=1.1.0

# 유틸리티
python-dotenv>=1.0.0  # 환경변수 관리

# 개발 도구 (선택사항)
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2  # 테스트용 HTTP 클라이언트

